backend/requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
environs[django]==9.3.2
phonenumbers==8.13.24
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
requests==2.31.0
geopy==2.4.0
rollbar==1.0.0
